+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- Add `validate_access_token` function to providers
+    - This can be used to verify the access token received from providers.
+
 ## [0.16.3] - 2023-09-28
 
 - Add Twitter provider for thirdparty login