+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [0.16.4] - 2023-10-05
+
 - Add `validate_access_token` function to providers
     - This can be used to verify the access token received from providers.
     - Implemented `validate_access_token` for the Github provider.